refactor: create new structure for endponts under data_bridges_knots (empty files as submodules)
