+++ data_bridges_knots/__init__.py
@@ -4,7 +4,7 @@
 Wrapper for DataBridges client.
 """
 
-from .client import DataBridgesShapes
+from .client import DataBridgesShapes, config_from_env
 from .labels import get_choice_labels, get_variable_labels, map_value_labels
 
 __all__ = [
@@ -13,4 +13,5 @@ __all__ = [
     "get_variable_labels",
     "get_choice_labels",
     "map_value_labels",
+    "config_from_env",
 ]
